added appointment migration

+++ data/migrations/20190428201604_client_table.js
@@ -16,8 +16,9 @@ exports.up = function(knex) {
             .notNullable()
             .unique();
         
+        // may be parsed as string type in javascript
         clients
-            .integer('phone', 10);
+            .bigInteger('phone', 10);
         
         clients
             .string('address', 256)
